update kelola kategori

+++ database/migrations/2024_07_11_092114_create_kategoris_table.php
@@ -15,8 +15,8 @@ return new class extends Migration {
             $table->id();
             $table->string('kategori');
             $table->text('keterangan');
-            $table->integer('range_min');
-            $table->integer('range_max');
+            $table->double('range_min');
+            $table->double('range_max');
             $table->foreignIdFor(Kriteria::class);
             $table->timestamps();
         });
